add validation error to clean() method, clean up,and set SUPERUSER_AUTH_ENFOREMENT to False
