+++ hub/admin/extend_user.py
@@ -15,7 +15,7 @@ from django.urls import reverse
 from django.utils import timezone
 from django.utils.safestring import mark_safe
 
-from hub.admin.validators import validate_superuser_auth
+from kobo.apps.accounts.mfa.models import MfaMethod
 from kobo.apps.accounts.validators import (
     USERNAME_MAX_LENGTH,
     USERNAME_INVALID_MESSAGE,
@@ -32,6 +32,15 @@ from kpi.models.asset import AssetDeploymentStatus
 from .filters import UserAdvancedSearchFilter
 from .mixins import AdvancedSearchMixin
 
+def validate_superuser_auth(obj) -> bool:
+    if (
+        obj.is_superuser
+        and config.SUPERUSER_AUTH_ENFORCEMENT
+        and obj.has_usable_password()
+        and not MfaMethod.objects.filter(user=obj, is_active=True).exists()
+    ):
+        return False
+    return True
 
 class UserChangeForm(DjangoUserChangeForm):
 
@@ -54,6 +63,11 @@ class UserChangeForm(DjangoUserChangeForm):
                 f'User is in <a href="{url}">trash</a> and cannot be reactivated'
                 f' from here.'
             ))
+        is_superuser = cleaned_data.get('is_superuser', False)
+        if is_superuser and not validate_superuser_auth(self.instance):
+            raise ValidationError(
+                "Superusers with a usable password must enable MFA."
+            )
 
         return cleaned_data
 
@@ -311,7 +325,3 @@ class ExtendedUserAdmin(AdvancedSearchMixin, UserAdmin):
             message += f'View <a href="{url}">trash.</a>'
 
         return mark_safe(message)
-
-    def save_model(self, request, obj, form, change):
-        validate_superuser_auth(obj)
-        super().save_model(request, obj, form, change)
